+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Change log
 
+**0.17.0**
+
+- Allow to quick pick pait of folders from recent compares
+- Allow to clear the recent compares list
+- New icons
+
 **0.16.0**
 
 Allow deleting file